move layout higher, fix adding v1 images
