+++ src/pkg/layout/component.go
