+++ src/pkg/layout/sbom.go
